backend/requirements.txt
# Backend dependencies (torch installed separately in Dockerfile)
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
voyageai==0.2.1
tokenizers==0.22.1
python-dotenv==1.0.0
pydantic==2.5.3
groq>=0.11.0
